Refactor commands into day command generator
Refactored the code so that the commands were generated by a function.
Since they are all really similar, it reduces the complexity when it
comes to reading the code. Now to add a new command, you only need to
add a single line to the `root.go` file. This is much easier to read
compared to creating a whole file that is essentially the same as other
days.