Refactored solutions into separate packages
Code structure was really loose to start with. There were a lot of
exported names in the solutions package, and most day functions were
modularized by a prefix of "day0X" on the function name. To make things
a little easier to read, and to make code easier to reference in the
language server, I've refactored the solutions into separate day
packages.

In addition to this change, I also moved the separate commands out to
their own files. It was starting to be difficult to understand how to
best add a new command with it's own custom inputs. To ease this, I've
essentially reverted my code to a state that a simple `cobra add`
command would add. This increases the amount of code, but now each
command is fully defined in it's own area.